+++ local_R/test_NW_monitors.R
@@ -12,7 +12,7 @@ monitor <- monitor_loadLatest() %>% monitor_subset(stateCodes = "WA")
 lons <- monitor$meta$longitude
 lats <- monitor$meta$latitude
 
-jon <- jon %>% addLocations(lons, lats, radius = 500)
+jon <- jon %>% addLocation(lons, lats, radius = 500)
 
 saveLocationTable(jon, "jon")
 
@@ -22,7 +22,7 @@ monitor <- monitor_loadLatest() %>% monitor_subset(stateCodes = c("WA","OR"))
 lons <- monitor$meta$longitude
 lats <- monitor$meta$latitude
 
-jon <- jon %>% addLocations(lons, lats, radius = 500)
+jon <- jon %>% addLocation(lons, lats, radius = 500)
 
 saveLocationTable(jon, "jon")
 
@@ -33,7 +33,7 @@ indices <- sort(sample(seq_along(monitor$meta$longitude), 10))
 lons <- monitor$meta$longitude[indices]
 lats <- monitor$meta$latitude[indices]
 
-jon <- jon %>% addLocations(lons, lats, radius = 500)
+jon <- jon %>% addLocation(lons, lats, radius = 500)
 
 saveLocationTable(jon, "jon")
 
@@ -43,7 +43,7 @@ monitor <- monitor_loadLatest() %>% monitor_subset(stateCodes = c("WA","OR","ID"
 lons <- monitor$meta$longitude
 lats <- monitor$meta$latitude
 
-jon <- jon %>% addLocations(lons, lats, radius = 500)
+jon <- jon %>% addLocation(lons, lats, radius = 500)
 
 saveLocationTable(jon, "jon")