+++ NEWS.md
@@ -1,3 +1,9 @@
+# Mazama LocationUtil 0.1.2
+
+* Renamed `addLocation()` to `addSingleLocation()`
+* Renamed `addLocations()` to `addLocation()`
+
+
 # Mazama LocationUtil 0.1.1
 
 * Removed `getLoations()`